requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis for queues
redis==5.0.1
celery==5.3.4

# Video processing
ffmpeg-python==0.2.0
moviepy==1.0.3

# Video downloading
yt-dlp==2023.11.16
instaloader==4.10.3
vk-api==11.9.9

# ML/AI for transcription
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1

# Web scraping & HTTP
aiohttp==3.9.1
requests==2.31.0
beautifulsoup4==4.12.2

# File handling
python-docx==1.1.0
webvtt-py==0.4.6

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# System monitoring
psutil==5.9.6

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1